Fix cassanderl application dependency.
When cassanderl is starting up, it needs to be dependent on `thrift`
and `cassandra_thrift`. None of these start anything but the
dependency needs to be there so the reltool can automatically figure
out the dependency. It also means errors will make the application
startup fail, which is what we want over a half-running system.