Attempt to fix some potential socket leak
When someone brutally kills a process using a thrift client connection,
it is possible for that socket to remain active even if used by anyone.

This patch makes it so that the dispcount callback module always tries
to close the socket when a resource is seen as dead before reallocating
a new onem in the hopes of getting rid of the leak altogether.